requirements.txt
flask
flask-cors
torch
torchvision
gdown
pillow